requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
googlesearch-python>=1.2.0
pandas>=2.2.0
lxml>=5.0.0
python-dotenv>=1.0.0
PyYAML>=6.0.1
